Refactor external function impls
