+++ src/lib.rs
@@ -1,6 +1,5 @@
 mod buffer;
 mod execute;
-mod externals;
 mod resolver;
 mod runtime;
